refactor mail merge sidebar
split SidebarContent in controller and view
add documentaiton
rename mail merge sidebar classes
fix initialization of TextDocumentController
